+++ HISTORY.rst
@@ -6,6 +6,9 @@ Next Release
 * Improve the biomass reaction finding logic.
 * Explicitly register custom markers `biomass`, `essentiality` and `growth`
   used for custom test parametrization in pytest.
+* Fix logic for `consistency` tests of production and consumption of 
+  metabolites with open bounds. It allows multiprocessing, currently relying on 
+  `cobra.Configuration` to select the number of processors.
 
 0.9.11 (2019-04-23)
 -------------------